* Fixed some integer type casts identified by Apple LLVM compiler.  * Fixed a potential memory leak in cstring_append with realloc identified by static code analysis.  * Changed xmlXpathFreeNodeList to xmlXPathFreeObject.  The former was not deallocating the nodes and leaking memory.  * Added some missing _destroy calls in the unit tests.  * Fixed some other misc memory leaks.
git-svn-id: https://atmos-c.googlecode.com/svn/branches/dev-atmos-c-2.1@88 cb61ccbc-2dfe-265f-a3f4-a677fceac7d0

+++ lib/atmos_service_info.c
@@ -52,7 +52,7 @@ void AtmosFilter_parse_service_info_response(RestFilter *self, RestClient *rest,
      * The document being in memory, it have no base per RFC 2396,
      * and the "noname.xml" argument will serve as its base.
      */
-    doc = xmlReadMemory(response->body, response->content_length,
+    doc = xmlReadMemory(response->body, (int)response->content_length,
             "noname.xml", NULL, 0);
     if (doc == NULL) {
         ATMOS_ERROR("Failed to parse error response: %s\n", response->body);
